Changed test for menuItemLogout and refactored package naming
