+++ app/src/androidTest/java/com/vinsonb/password/manager/kotlin/fragments/CreateLoginFragmentTest.kt
@@ -15,7 +15,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.vinsonb.password.manager.kotlin.R
-import com.vinsonb.password.manager.kotlin.matchers.TextInputLayoutMatchers.withTextInputLayoutErrorText
+import com.vinsonb.password.manager.kotlin.utilities.TextInputLayoutMatchers.withTextInputLayoutErrorText
 import org.hamcrest.CoreMatchers.not
 import org.junit.After
 import org.junit.Assert
