+++ app/src/androidTest/java/com/vinsonb/password/manager/kotlin/utilities/RecyclerViewMatchers.kt
@@ -1,4 +1,4 @@
-package com.vinsonb.password.manager.kotlin.matchers
+package com.vinsonb.password.manager.kotlin.utilities
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
